Separate `TensorMeshHierarchy` and iterators.
