+++ src/TensorMeshHierarchyIteration.cpp
@@ -1,4 +1,4 @@
-#include "TensorMeshHierarchy.hpp"
+#include "TensorMeshHierarchyIteration.hpp"
 
 #include <stdexcept>
